feat: basic framework for the Binius bindings (#28)
Creating bindings to Binius is tricky because the main types don't implement
`Clone` and their methods work on the basis of mutation.

The solution proposed on this patch hides the interior mutability of the Lean
object in a private function, used in a stateful monad, as if the state were
being overwritten.