feat: bindings for Binius witness generation (#46)
This patch not only adds support for witness generation in binius, but also does
a big refactor on our FFI structure.

First, we should not allocate Lean data from Rust because Lean performs a
(compile) flag-based allocation of Lean objects via `lean_alloc_object`. The
Rust API for Lean objects should be used for read-only purposes from now on.

Lower level APIs for `UInt64`, `USize` and `UInt128` are now purely in C. The
API is now unified in a single `Unsigned.lean` file.

This patch also implements serde validation for the `Boundary` type, with a
property test asserting that converting from a Lean object to a Rust object, in
Rust, arrives at the correct `boundary: Boundary`.

Lastly, add the `never_extract` tag for functions that perform side-effects or
functions that return objects that work on the basis of side-effects. This is
important for Lean not to perform simplifications that assume purity/idempotence
of implementations.