chore: prevent memory leak of ditched linear objects
Before this patch, marking a linear object as outdated would set `object_ref` to
`NULL`. Then, if `allow_finalizer` is `true`, the finalizer will be called on
`NULL`, which will be a NOOP.

This patch prevents the loss of the object reference by using a `bool` to track
outdated linear objects. Then the finalizer will not be called on `NULL` and the
memory won't leak.