feat: Archon
The `archon` crate is an abstraction layer on top of Binius that
enables the construction of self-contained circuit modules, whose
respective witnesses can be populated independently.

Building a circuit module doesn't require knowledge about witness
size, which qualifies a set of circuit modules as a stable protocol
contract between provers and verifiers.

When compiling witness modules, empty ones are skipped, meaning
the corresponding circuit module will be considered deactivated.

Compilation is the process of aggregating multiple modules into a
single entity. Witness modules are compiled into a `Witness` and
circuit modules are compiled into a `ConstraintSystem` (from the
`binius` crate). But since `ConstraintSystem` needs the sizes of
the populated oracles, compiling circuit modules requires this
information, which is present on the compiled `Witness`.

The `protocol.rs` file puts all this together, providing the
functions `validate_witness`, `prove` and `verify`.

This patch doesn't include the following planned features:
* Automatic versioning of canonicalized circuit modules
* Automatic populating logic for oracles that can be computed