chore: redefine `G` in Lean
Use a subtype whose inner value is an `UInt64`, constrained to be
smaller than the size of the Goldilocks field.

This definition of `G` entails a much better memory layout for
interfacing with Rust, as it is not conditioned on the size of the
number being wrapped like `Nat` is.