feat: Aiur gadgets infra and a byte gadget (#202)
This patch defines the `AiurGadget` trait, which enables gadgets to
be plugged into Aiur machinery.

As a example, the `Bytes1` gadget was added, arming Aiur with
bit decomposition, left and right shifting of bytes.

These changes are propagated all the way up to the Lean frontend,
with a test showcasing the prove/verify pipeline succeeding.