feat: `assert_eq!` Aiur operation
`assert_eq!` in Aiur creates constraints that require equalities
of the corresponding circuit values.

Notably, it's consistent for pointers if they're of the same type
(they point to data of the same width). That is, if `ptr1: &T` and
`ptr2: &T`, then `assert_eq!(ptr1, ptr2)` enforces that the data
they're pointing at must be equal.