feat: Aiur `Unit` type and `assert_eq!` operation (#215)
* feat: introduce the "Unit" type on Aiur frontend

Add a `Term.unit` and its corresponding `Typ.unit` to Aiur's frontned.
This term has size zero. It's meant to be used as the term returned
by functions that "return nothing". For example, a function that simply
writes data to the `IOBuffer`.

* feat: `assert_eq!` Aiur operation

`assert_eq!` in Aiur creates constraints that require equalities
of the corresponding circuit values.

Notably, it's consistent for pointers if they're of the same type
(they point to data of the same width). That is, if `ptr1: &T` and
`ptr2: &T`, then `assert_eq!(ptr1, ptr2)` enforces that the data
they're pointing at must be equal.