feat: `#[unconstrained]` tag for Aiur functions
An unconstrained function has no constraints or lookups. Therefore
it can accept a dummy single-row trace filled with zeros.

Extra care must be taken when calling an unconstrained function,
as such function is not going to pull any claim. So the caller can't
push the claim either.

Furthermore, an unconstrained call must propagate "unconstrainedness".
That is, for example, if an unconstrained function calls a properly
constrained function, any subconsequent transitive lookup performed
through this call must not be accounted for.