Start `Ixon` serde in Aiur (#268)
This patch implements (and changes) enough of the code base as a PoC of being able to provably bootstrap the precise Ixon whose hash is stated as an argument of the claim (as input of the Aiur entrypoint call).

This is done by reading the bytes from the `IOBuffer`, deserializing it and then serializing and hashing the respective bytes. Deserialization followed by serialization was the chosen solution because serialization is, in principle, cheaper. Thus the deserialization can be tagged as `#[unconstrained]` for extra speed.

Furthermore, the unconstrained tag is now more rigorous: an unconstrained function can no longer call a constrained function.

As an extra, the IxVM code has been split into multiple smaller Aiur toplevels to enable parallel elaboration by Lean.

---------

Co-authored-by: Gabriel Barreto <gabriel.aquino.barreto@gmail.com>